+++ packages/streams_provider/lib/src/extensions/stream_subscription.dart
@@ -1,9 +1,0 @@
-import 'dart:async';
-
-import '../streams_dispose_bag.dart';
-
-extension StreamSubscriptionExtensions<T> on StreamSubscription<T> {
-  void disposed({required StreamsDisposeBag by}) {
-    by.add(this);
-  }
-}