+++ packages/streams_provider/lib/streams_provider.dart
@@ -1,15 +1,12 @@
 library stream_provider;
 
-export 'src/mutable_value_stream.dart';
 export 'src/streams_selector.dart';
 export 'src/extensions/rx_dart.dart';
 export 'src/streams_providable.dart';
 export 'src/streams_provider.dart';
 export 'src/streams_listener.dart';
 export 'src/streams_listenable.dart';
-export 'src/streams_dispose_bag.dart';
 export 'package:rxdart/rxdart.dart';
 export 'src/matchers.dart';
 export 'src/extensions/text_editing_controller.dart';
 export 'src/multi_streams_listener.dart';
-export 'src/extensions/stream_subscription.dart';